requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
celery==5.3.4
kombu==5.3.4
amqp==5.2.0
python-dotenv==1.0.0
pandas==2.1.4
supabase==2.3.0
gunicorn==21.2.0
requests==2.31.0

